Card modal to show more info onf selected card. Code cleanup. New JS file to hold styles for card elements.
Card popup calls fetch method on new enpoint - endpoint retrieves all info on a specific card using the cards ID. Card popup uses the same styles as the standalone cards seen in the grid layout. To simplify this a new JS file was made to hold the styles. These styles will be referenced whenever styling a card is needed.